+++ service/controller/control-plane/controller.go
@@ -1,17 +1,17 @@
-package legacy
+package controlplane
 
 import (
 	promclient "github.com/coreos/prometheus-operator/pkg/client/versioned"
-	"github.com/giantswarm/apiextensions/v2/pkg/apis/provider/v1alpha1"
 	"github.com/giantswarm/k8sclient/v4/pkg/k8sclient"
 	"github.com/giantswarm/microerror"
 	"github.com/giantswarm/micrologger"
 	"github.com/giantswarm/operatorkit/v2/pkg/controller"
 	"github.com/giantswarm/operatorkit/v2/pkg/resource"
+	v1 "k8s.io/api/core/v1"
+	"k8s.io/apimachinery/pkg/labels"
 	"k8s.io/apimachinery/pkg/runtime"
 
 	"github.com/giantswarm/prometheus-meta-operator/pkg/project"
-	controllerresource "github.com/giantswarm/prometheus-meta-operator/service/controller/resource"
 )
 
 type ControllerConfig struct {
@@ -33,42 +33,33 @@ func NewController(config ControllerConfig) (*Controller, error) {
 
 	var resources []resource.Interface
 	{
-		c := controllerresource.Config(config)
+		c := resourcesConfig(config)
 
-		resources, err = controllerresource.New(c)
+		resources, err = newResources(c)
 		if err != nil {
 			return nil, microerror.Mask(err)
 		}
 	}
 
+	var selector controller.Selector
+	{
+		set := labels.Set{
+			"component": "apiserver",
+		}
+		selector = set.AsSelector()
+	}
+
 	var operatorkitController *controller.Controller
 	{
-		var runtimeFunc func() runtime.Object
-		{
-			switch config.Provider {
-			case "aws":
-				runtimeFunc = func() runtime.Object {
-					return new(v1alpha1.AWSConfig)
-				}
-			case "azure":
-				runtimeFunc = func() runtime.Object {
-					return new(v1alpha1.AzureConfig)
-				}
-			case "kvm":
-				runtimeFunc = func() runtime.Object {
-					return new(v1alpha1.KVMConfig)
-				}
-			default:
-				return nil, microerror.Maskf(invalidProviderError, "provider: %q", config.Provider)
-			}
-		}
-
 		c := controller.Config{
-			K8sClient:            config.K8sClient,
-			Logger:               config.Logger,
-			Name:                 project.Name() + "-legacy-controller",
-			NewRuntimeObjectFunc: runtimeFunc,
-			Resources:            resources,
+			K8sClient: config.K8sClient,
+			Logger:    config.Logger,
+			Name:      project.Name() + "-control-plane-controller",
+			NewRuntimeObjectFunc: func() runtime.Object {
+				return new(v1.Service)
+			},
+			Resources: resources,
+			Selector:  selector,
 		}
 
 		operatorkitController, err = controller.New(c)